Add multipication of count property by numeric key
This commit resolves #288
For example, if you type `3J`, then go to the third tab to the left.
This feauter is available for tabs.prev, tabs.next, scroll.vertically,
scroll.horizonally, and scroll.pages because these operation have count property.
If you set count property of j to 5 and type `2j` then scroll distance is
10 times longer than default.